+++ src/fmu/dataio/providers/__init__.py
@@ -1,6 +1,6 @@
 from ._filedata import FileDataProvider, SharePathConstructor
 from ._fmu import FmuProvider
-from .objectdata._base import ObjectDataProvider, UnsetData
+from .objectdata._base import ObjectDataProvider
 from .objectdata._provider import objectdata_provider_factory
 
 __all__ = [
@@ -8,6 +8,5 @@ __all__ = [
     "SharePathConstructor",
     "FmuProvider",
     "ObjectDataProvider",
-    "UnsetData",
     "objectdata_provider_factory",
 ]
